* Updated Makefile to always include alt_ios_gen.c * Changed default font to DejaVu (Russian glyphs are removed for now!) * Changed list to include a lookup in custom_titles * Fixed problem with gamesound when exiting submenus like synopsis or settings
git-svn-id: https://wiiflow-advanced.googlecode.com/svn/trunk@368 43c7226f-5572-e80b-0314-b0280708705f

+++ source/gui/WiiTDB.hpp
@@ -52,7 +52,7 @@ typedef struct _GameXMLInfo
 	safe_vector<string> WifiFeatures;
 	int Players;
 	safe_vector<Accessoire> Accessoires;
-	long CaseColor;
+	int CaseColor;
 
 } GameXMLInfo;
